+++ src/main/java/org/wwu/bpm/gta/creditscore/CheckPayrollInformation.java
@@ -9,7 +9,7 @@ public class CheckPayrollInformation extends ServletProcessApplication  implemen
 	@Override
 	public void execute(DelegateExecution execution) throws Exception {
 		// generates random variable in value range from 0 to 10 that represents the ability to pay, based on applicants payroll
-				int PayrollScore = (int)(Math.random() * 1);	
+				double PayrollScore = (double)(Math.random() * 1);	
 				execution.setVariable("PayrollScore", PayrollScore);
 		
 	}
